latest updates:
i have saved all my changes to the map branch. i'll leave it to James the Github guru (or anyone else competent in Github) to merge to master as they see fit

I have gotten all the links working and the main page opens to the map.

One thing to note is when a user wants to go back to the map from another page, they need to click the "refresh button" as it seems that the only way to make sure the map loads each time from a relative link is to fully reload it